# Optimize StagedMetadatas conversion