--- a/src/metrics/metric/aggregated/types_test.go
+++ b/src/metrics/metric/aggregated/types_test.go
@@ -38,6 +38,7 @@ import (
 	"github.com/m3db/m3/src/metrics/transformation"
 	xtime "github.com/m3db/m3/src/x/time"
 
+	"github.com/google/go-cmp/cmp"
 	"github.com/stretchr/testify/require"
 )
 
@@ -365,13 +366,18 @@ func TestForwardedMetricWithMetadataFromProto(t *testing.T) {
 	}
 
 	var res ForwardedMetricWithMetadata
+	comparer := cmp.Comparer(spComparer)
 	for _, input := range inputs {
 		require.NoError(t, res.FromProto(&input.data))
 		expected := ForwardedMetricWithMetadata{
 			ForwardedMetric: input.expectedMetric,
 			ForwardMetadata: input.expectedMetadata,
 		}
-		require.Equal(t, expected, res)
+
+		if !cmp.Equal(expected, res, comparer) {
+			t.Log(cmp.Diff(expected, res, comparer))
+			t.Fail()
+		}
 	}
 }
 
@@ -416,6 +422,7 @@ func TestForwardedMetricWithMetadataRoundtrip(t *testing.T) {
 		res ForwardedMetricWithMetadata
 		pb  metricpb.ForwardedMetricWithMetadata
 	)
+	comparer := cmp.Comparer(spComparer)
 	for _, input := range inputs {
 		data := ForwardedMetricWithMetadata{
 			ForwardedMetric: input.metric,
@@ -423,6 +430,13 @@ func TestForwardedMetricWithMetadataRoundtrip(t *testing.T) {
 		}
 		require.NoError(t, data.ToProto(&pb))
 		require.NoError(t, res.FromProto(&pb))
-		require.Equal(t, data, res)
+		if !cmp.Equal(data, res, comparer) {
+			t.Log(cmp.Diff(data, res, comparer))
+			t.Fail()
+		}
 	}
+}
+
+func spComparer(x, y policy.StoragePolicy) bool {
+	return x.Equivalent(y)
 }