… src/metrics/encoding/protobuf/unaggregated_encoder_test.go
… src/metrics/metadata/metadata_benchmark_test.go
… src/metrics/metadata/metadata_test.go
… src/metrics/pipeline/applied/type_test.go